CSES Problems of 1D dp space optimized and overflow safed versions

+++ leetcode/BinarySearchOnAnswers/GasStation/solution.java
@@ -90,7 +90,7 @@ public class solution {
             numbersBtween--;
            }
 
-           cnt += (int) numbersBtween;
+           cnt += numbersBtween;
         }
         return cnt;
     }
